chore: rename to floyd warshall paths and update logic
