+++ internal/transducer/booking_test.go
@@ -28,11 +28,11 @@ func TestRehearseBookingTransducer(t *testing.T) {
 	}
 }
 
-func TestBookingShortestPaths(t *testing.T) {
+func TestBookingGetFloydWarshallPaths(t *testing.T) {
 	characterConfig := CreateConfig().SetState(Idle)
 	characterTransducer := NewBookingTransducer(characterConfig)
 
-	paths, edges := characterTransducer.GetShortestPaths()
+	paths, edges := characterTransducer.GetFloydWarshallPaths()
 	currentState := graph.Vertex(Idle)
 
 	for _, path := range paths {
